Make agile swimlane badge readable
The style itself is somewhat bad to recognize where it actually matches.
The previous behaviour had it that the swimlane badge background which
contains the issue ID link (f. i. `Epic #1234`) has the same color as the
text color. Essentially it was an unreadable blob of color.

The style makes the issue ID link readable again. This works on open and also an closed swimlanes.