Fix sorting behavior on enumeration page
Sorting a list item on administration/enumerations page causes for the
tables at second and third position on the page an unexpected behavior
when trying to sort an element by drag & drop.

That is, the distance between the dragged element and its placeholder
increases that much that it was difficult to position the list item as
wished.